// angular-src/src/app/interfaces/usuario.interface.ts
export interface Usuario {
  _id?: string;
  nombre?: string;
  apellido?: string;
  edad?: number;
  sexo?: string;
  peso?: number | string;
  altura?: number | string;
  doctor?: string;
  doctorNombre?: string;
  doctorId?: string;
  correo?: string;
  contrasena?: string;
  usuario?: string;
  nombreCompleto?: string;
  recaptcha?: string;
}